Start work at handling previous moves
